move tests to test/tests folder
