+++ test/tests/action/clickoutside/Component.svelte
