+++ test/tests/store/resettable/resettable.test.ts
@@ -1,6 +1,6 @@
 import { get } from 'svelte/store';
 import { describe, expect, it } from 'vitest';
-import { resettable } from '../../../src/lib/store/resettable';
+import { resettable } from '../../../../src/lib/store/resettable';
 
 describe('Resettable stores', () => {
 	it('should have a reset function', () => {
